pyproject.toml
[tool.poetry]
name = "GTBO"
version = "0.1.0"
description = ""
authors = ["Anonymous authors"]
readme = "README.md"
packages = [{ include = "gtbo" }]

[tool.poetry.dependencies]
python = "^3.9.0"
numpy = "*"
torch = "<=2.0.0"
botorch = "^0.9.2"
gin-config = "^0.5.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pdoc = "^14.1.0"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
